extract commands + add command shortcuts + cleanup
